Multiple aspects and generation of sets
Allows to select from a combination of two or more aspects in the csv file. The new approach now generates a single DataFrame containing the classes (labels) of emtions and all aspects to be trained. Also generates training and test sets, based on a sampled (randomized) version of the DataFrame.